ensure that libardour_cp and libardour SMPTE structs are same type, hackishly
git-svn-id: svn://localhost/trunk/ardour2@525 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/ardour/types.h
@@ -106,14 +106,14 @@ namespace ARDOUR {
 		smpte_wrap_minutes,
 		smpte_wrap_hours
 	};
-  
+
 	struct SMPTE_Time {
 		bool negative;
-		long hours;
-		long minutes;
-		long seconds;
-		long frames;
-		long subframes; // mostly not used
+		uint32_t hours;
+		uint32_t minutes;
+		uint32_t seconds;
+		uint32_t frames;
+		uint32_t subframes; // mostly not used
 
 		SMPTE_Time() {
 			negative = false;
